requirements-test.txt
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
mock>=4.0.3
requests-mock>=1.10.0
